requirements.txt
alembic==1.3.2
amqp==2.5.2
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
billiard==3.6.1.0
blinker==1.4
celery==4.4.0
Click==7.0
colorama==0.4.3
coverage==5.0.1
Flask==1.1.1
Flask-Caching==1.8.0
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
kombu==4.6.7
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
pycodestyle==2.5.0
pylint==2.4.4
pyrsistent==0.15.6
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
redis==3.3.11
six==1.13.0
SQLAlchemy==1.3.12
typed-ast==1.4.0
vine==1.3.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
